+++ CHANGES.md
@@ -3,6 +3,12 @@
 Fedify changelog
 ================
 
+Version 0.7.0
+-------------
+
+To be released.
+
+
 Version 0.6.0
 -------------